+++ src/components.d.ts
@@ -5,7 +5,7 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { ButtonSizeEnum, ButtonVariantBean, ComboItemBean, DividerOrientation, DividerSize, HeaderUserData, InputStatusBean, InputTypeBean, MenuItem, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, StatusTagStatus, TableHeaderSize, ThemeVariant, ThemeVariantBean } from "./beans";
+import { ButtonSizeEnum, ButtonVariantBean, CardVariants, ComboItemBean, DividerOrientation, DividerSize, HeaderUserData, InputStatusBean, InputTypeBean, MenuItem, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, StatusTagStatus, TableHeaderSize, TabOrientationBean, TabSizeBean, ThemeVariant, ThemeVariantBean } from "./beans";
 import { ListItemBean } from "./beans/index.js";
 import { LicenseTypeEnum, MenuItem as MenuItem1, TooltipPosition } from "./beans/index";
 import { ZTypographyLevels } from "./components/typography/z-typography/index";
@@ -153,6 +153,20 @@ export namespace Components {
         "sortlabeldesc"?: string;
     }
     interface ZCandybar {
+    }
+    interface ZCard {
+        /**
+          * Enable click interactions on the card. Default: false
+         */
+        "clickable": boolean;
+        /**
+          * Name of the icon to place over the image cover
+         */
+        "coverIcon": string;
+        /**
+          * Card variant. Can be one of "text", "border", "shadow", "overlay". Leave it undefined for the default card.
+         */
+        "variant": CardVariants;
     }
     interface ZChip {
         "boldtext"?: number;
@@ -748,6 +762,20 @@ export namespace Components {
          */
         "listdata": string;
     }
+    interface ZNavigationTab {
+        "disabled"?: boolean;
+        "selected"?: boolean;
+    }
+    interface ZNavigationTabs {
+        /**
+          * Available orientation: `horizontal` and `vertical`. Defaults to `horizontal`.
+         */
+        "orientation"?: TabOrientationBean;
+        /**
+          * Available sizes: `big` and `small`. Defaults to `big`.
+         */
+        "size"?: TabSizeBean;
+    }
     interface ZPaginationBar {
         /**
           * current displayed page (mutable)
@@ -878,6 +906,12 @@ export namespace Components {
          */
         "showButton"?: boolean;
     }
+    interface ZRegistroTableExpandedRow {
+        /**
+          * Number table column plus 1 for the expand button
+         */
+        "colSpan": number;
+    }
     interface ZRegistroTableFooter {
     }
     interface ZRegistroTableHead {
@@ -900,7 +934,11 @@ export namespace Components {
          */
         "sortable"?: boolean;
     }
+    interface ZRegistroTableHeaderRow {
+        "expandable"?: boolean;
+    }
     interface ZRegistroTableRow {
+        "expandable"?: boolean;
     }
     interface ZRegistroTableStickyFooter {
     }
@@ -1113,6 +1151,12 @@ declare global {
         prototype: HTMLZCandybarElement;
         new (): HTMLZCandybarElement;
     };
+    interface HTMLZCardElement extends Components.ZCard, HTMLStencilElement {
+    }
+    var HTMLZCardElement: {
+        prototype: HTMLZCardElement;
+        new (): HTMLZCardElement;
+    };
     interface HTMLZChipElement extends Components.ZChip, HTMLStencilElement {
     }
     var HTMLZChipElement: {
@@ -1311,6 +1355,18 @@ declare global {
         prototype: HTMLZMyzCardListElement;
         new (): HTMLZMyzCardListElement;
     };
+    interface HTMLZNavigationTabElement extends Components.ZNavigationTab, HTMLStencilElement {
+    }
+    var HTMLZNavigationTabElement: {
+        prototype: HTMLZNavigationTabElement;
+        new (): HTMLZNavigationTabElement;
+    };
+    interface HTMLZNavigationTabsElement extends Components.ZNavigationTabs, HTMLStencilElement {
+    }
+    var HTMLZNavigationTabsElement: {
+        prototype: HTMLZNavigationTabsElement;
+        new (): HTMLZNavigationTabsElement;
+    };
     interface HTMLZPaginationBarElement extends Components.ZPaginationBar, HTMLStencilElement {
     }
     var HTMLZPaginationBarElement: {
@@ -1353,6 +1409,12 @@ declare global {
         prototype: HTMLZRegistroTableCellElement;
         new (): HTMLZRegistroTableCellElement;
     };
+    interface HTMLZRegistroTableExpandedRowElement extends Components.ZRegistroTableExpandedRow, HTMLStencilElement {
+    }
+    var HTMLZRegistroTableExpandedRowElement: {
+        prototype: HTMLZRegistroTableExpandedRowElement;
+        new (): HTMLZRegistroTableExpandedRowElement;
+    };
     interface HTMLZRegistroTableFooterElement extends Components.ZRegistroTableFooter, HTMLStencilElement {
     }
     var HTMLZRegistroTableFooterElement: {
@@ -1370,6 +1432,12 @@ declare global {
     var HTMLZRegistroTableHeaderElement: {
         prototype: HTMLZRegistroTableHeaderElement;
         new (): HTMLZRegistroTableHeaderElement;
+    };
+    interface HTMLZRegistroTableHeaderRowElement extends Components.ZRegistroTableHeaderRow, HTMLStencilElement {
+    }
+    var HTMLZRegistroTableHeaderRowElement: {
+        prototype: HTMLZRegistroTableHeaderRowElement;
+        new (): HTMLZRegistroTableHeaderRowElement;
     };
     interface HTMLZRegistroTableRowElement extends Components.ZRegistroTableRow, HTMLStencilElement {
     }
@@ -1440,6 +1508,7 @@ declare global {
         "z-button-filter": HTMLZButtonFilterElement;
         "z-button-sort": HTMLZButtonSortElement;
         "z-candybar": HTMLZCandybarElement;
+        "z-card": HTMLZCardElement;
         "z-chip": HTMLZChipElement;
         "z-combobox": HTMLZComboboxElement;
         "z-cookiebar": HTMLZCookiebarElement;
@@ -1473,6 +1542,8 @@ declare global {
         "z-myz-card-header": HTMLZMyzCardHeaderElement;
         "z-myz-card-icon": HTMLZMyzCardIconElement;
         "z-myz-card-list": HTMLZMyzCardListElement;
+        "z-navigation-tab": HTMLZNavigationTabElement;
+        "z-navigation-tabs": HTMLZNavigationTabsElement;
         "z-pagination-bar": HTMLZPaginationBarElement;
         "z-pagination-page": HTMLZPaginationPageElement;
         "z-panel-elem": HTMLZPanelElemElement;
@@ -1480,9 +1551,11 @@ declare global {
         "z-registro-table": HTMLZRegistroTableElement;
         "z-registro-table-body": HTMLZRegistroTableBodyElement;
         "z-registro-table-cell": HTMLZRegistroTableCellElement;
+        "z-registro-table-expanded-row": HTMLZRegistroTableExpandedRowElement;
         "z-registro-table-footer": HTMLZRegistroTableFooterElement;
         "z-registro-table-head": HTMLZRegistroTableHeadElement;
         "z-registro-table-header": HTMLZRegistroTableHeaderElement;
+        "z-registro-table-header-row": HTMLZRegistroTableHeaderRowElement;
         "z-registro-table-row": HTMLZRegistroTableRowElement;
         "z-registro-table-sticky-footer": HTMLZRegistroTableStickyFooterElement;
         "z-select": HTMLZSelectElement;
@@ -1651,6 +1724,24 @@ declare namespace LocalJSX {
         "sortlabeldesc"?: string;
     }
     interface ZCandybar {
+    }
+    interface ZCard {
+        /**
+          * Enable click interactions on the card. Default: false
+         */
+        "clickable"?: boolean;
+        /**
+          * Name of the icon to place over the image cover
+         */
+        "coverIcon"?: string;
+        /**
+          * Card click event. Fired only if `clickable` is true.
+         */
+        "onCardClicked"?: (event: CustomEvent<any>) => void;
+        /**
+          * Card variant. Can be one of "text", "border", "shadow", "overlay". Leave it undefined for the default card.
+         */
+        "variant"?: CardVariants;
     }
     interface ZChip {
         "boldtext"?: number;
@@ -2318,6 +2409,21 @@ declare namespace LocalJSX {
          */
         "listdata"?: string;
     }
+    interface ZNavigationTab {
+        "disabled"?: boolean;
+        "onSelected"?: (event: CustomEvent<any>) => void;
+        "selected"?: boolean;
+    }
+    interface ZNavigationTabs {
+        /**
+          * Available orientation: `horizontal` and `vertical`. Defaults to `horizontal`.
+         */
+        "orientation"?: TabOrientationBean;
+        /**
+          * Available sizes: `big` and `small`. Defaults to `big`.
+         */
+        "size"?: TabSizeBean;
+    }
     interface ZPaginationBar {
         /**
           * current displayed page (mutable)
@@ -2460,6 +2566,12 @@ declare namespace LocalJSX {
          */
         "showButton"?: boolean;
     }
+    interface ZRegistroTableExpandedRow {
+        /**
+          * Number table column plus 1 for the expand button
+         */
+        "colSpan"?: number;
+    }
     interface ZRegistroTableFooter {
     }
     interface ZRegistroTableHead {
@@ -2486,7 +2598,12 @@ declare namespace LocalJSX {
          */
         "sortable"?: boolean;
     }
+    interface ZRegistroTableHeaderRow {
+        "expandable"?: boolean;
+    }
     interface ZRegistroTableRow {
+        "expandable"?: boolean;
+        "onExpand"?: (event: CustomEvent<any>) => void;
     }
     interface ZRegistroTableStickyFooter {
     }
@@ -2662,6 +2779,7 @@ declare namespace LocalJSX {
         "z-button-filter": ZButtonFilter;
         "z-button-sort": ZButtonSort;
         "z-candybar": ZCandybar;
+        "z-card": ZCard;
         "z-chip": ZChip;
         "z-combobox": ZCombobox;
         "z-cookiebar": ZCookiebar;
@@ -2695,6 +2813,8 @@ declare namespace LocalJSX {
         "z-myz-card-header": ZMyzCardHeader;
         "z-myz-card-icon": ZMyzCardIcon;
         "z-myz-card-list": ZMyzCardList;
+        "z-navigation-tab": ZNavigationTab;
+        "z-navigation-tabs": ZNavigationTabs;
         "z-pagination-bar": ZPaginationBar;
         "z-pagination-page": ZPaginationPage;
         "z-panel-elem": ZPanelElem;
@@ -2702,9 +2822,11 @@ declare namespace LocalJSX {
         "z-registro-table": ZRegistroTable;
         "z-registro-table-body": ZRegistroTableBody;
         "z-registro-table-cell": ZRegistroTableCell;
+        "z-registro-table-expanded-row": ZRegistroTableExpandedRow;
         "z-registro-table-footer": ZRegistroTableFooter;
         "z-registro-table-head": ZRegistroTableHead;
         "z-registro-table-header": ZRegistroTableHeader;
+        "z-registro-table-header-row": ZRegistroTableHeaderRow;
         "z-registro-table-row": ZRegistroTableRow;
         "z-registro-table-sticky-footer": ZRegistroTableStickyFooter;
         "z-select": ZSelect;
@@ -2729,6 +2851,7 @@ declare module "@stencil/core" {
             "z-button-filter": LocalJSX.ZButtonFilter & JSXBase.HTMLAttributes<HTMLZButtonFilterElement>;
             "z-button-sort": LocalJSX.ZButtonSort & JSXBase.HTMLAttributes<HTMLZButtonSortElement>;
             "z-candybar": LocalJSX.ZCandybar & JSXBase.HTMLAttributes<HTMLZCandybarElement>;
+            "z-card": LocalJSX.ZCard & JSXBase.HTMLAttributes<HTMLZCardElement>;
             "z-chip": LocalJSX.ZChip & JSXBase.HTMLAttributes<HTMLZChipElement>;
             "z-combobox": LocalJSX.ZCombobox & JSXBase.HTMLAttributes<HTMLZComboboxElement>;
             "z-cookiebar": LocalJSX.ZCookiebar & JSXBase.HTMLAttributes<HTMLZCookiebarElement>;
@@ -2762,6 +2885,8 @@ declare module "@stencil/core" {
             "z-myz-card-header": LocalJSX.ZMyzCardHeader & JSXBase.HTMLAttributes<HTMLZMyzCardHeaderElement>;
             "z-myz-card-icon": LocalJSX.ZMyzCardIcon & JSXBase.HTMLAttributes<HTMLZMyzCardIconElement>;
             "z-myz-card-list": LocalJSX.ZMyzCardList & JSXBase.HTMLAttributes<HTMLZMyzCardListElement>;
+            "z-navigation-tab": LocalJSX.ZNavigationTab & JSXBase.HTMLAttributes<HTMLZNavigationTabElement>;
+            "z-navigation-tabs": LocalJSX.ZNavigationTabs & JSXBase.HTMLAttributes<HTMLZNavigationTabsElement>;
             "z-pagination-bar": LocalJSX.ZPaginationBar & JSXBase.HTMLAttributes<HTMLZPaginationBarElement>;
             "z-pagination-page": LocalJSX.ZPaginationPage & JSXBase.HTMLAttributes<HTMLZPaginationPageElement>;
             "z-panel-elem": LocalJSX.ZPanelElem & JSXBase.HTMLAttributes<HTMLZPanelElemElement>;
@@ -2769,9 +2894,11 @@ declare module "@stencil/core" {
             "z-registro-table": LocalJSX.ZRegistroTable & JSXBase.HTMLAttributes<HTMLZRegistroTableElement>;
             "z-registro-table-body": LocalJSX.ZRegistroTableBody & JSXBase.HTMLAttributes<HTMLZRegistroTableBodyElement>;
             "z-registro-table-cell": LocalJSX.ZRegistroTableCell & JSXBase.HTMLAttributes<HTMLZRegistroTableCellElement>;
+            "z-registro-table-expanded-row": LocalJSX.ZRegistroTableExpandedRow & JSXBase.HTMLAttributes<HTMLZRegistroTableExpandedRowElement>;
             "z-registro-table-footer": LocalJSX.ZRegistroTableFooter & JSXBase.HTMLAttributes<HTMLZRegistroTableFooterElement>;
             "z-registro-table-head": LocalJSX.ZRegistroTableHead & JSXBase.HTMLAttributes<HTMLZRegistroTableHeadElement>;
             "z-registro-table-header": LocalJSX.ZRegistroTableHeader & JSXBase.HTMLAttributes<HTMLZRegistroTableHeaderElement>;
+            "z-registro-table-header-row": LocalJSX.ZRegistroTableHeaderRow & JSXBase.HTMLAttributes<HTMLZRegistroTableHeaderRowElement>;
             "z-registro-table-row": LocalJSX.ZRegistroTableRow & JSXBase.HTMLAttributes<HTMLZRegistroTableRowElement>;
             "z-registro-table-sticky-footer": LocalJSX.ZRegistroTableStickyFooter & JSXBase.HTMLAttributes<HTMLZRegistroTableStickyFooterElement>;
             "z-select": LocalJSX.ZSelect & JSXBase.HTMLAttributes<HTMLZSelectElement>;